On branch master
 Changes to be committed:
		modified:   pynder/models/user.py
		modified:   pynder/session.py

Added docstrings to many of the functions.

Also made modifications so that many functions now return
a 'zero value' of whatever object you would expect that function
to return so as to make things crash less often.

Most of the changes were done to the User model.

The following properties were added:
	mentions_snapchat
	mentions_kik

Also a dict() method was added to User to get an easy dictionary
representation of the object.

Unfortunately, the nosetests fail with this version of the library,
but I suspect it has something to do with how the tests are written
themselves since the libary seems to be working as expected when not
run in the test runner. I'm not familiar with tox, nosetest, or vcrpy
so I'll try and get up to speed on those libraries and see if I can
fix whatever is going on.